water_flowmeter was initially detected as a string
Later it's a struct, which skippr detected and handled. However, more fields were added to that struct, which it seems can't be evolved.

Note the new is_fitted isn’t present in the existing metadata which means the field wasn’t present in the data when we discovered and evolved the complex field type (previously a string).

Unfortunately, skippr didn’t have support to then further evolve the complex field as new subfields were added or types changed.